Allow for 64 tracks per media type

+++ vod/media_format.h
@@ -159,7 +159,7 @@ typedef struct {
 } media_range_t;
 
 typedef struct {
-	uint32_t* required_tracks_mask;
+	uint64_t* required_tracks_mask;
 	uint8_t* langs_mask;
 	uint32_t clip_from;
 	uint32_t clip_to;
